Lobby refactoring and changes
