+++ launcher/lobby/lobby_moc.ui
@@ -84,30 +84,6 @@
      </property>
      <widget class="QWidget" name="sessionsPage">
       <layout class="QGridLayout" name="gridLayout_2">
-       <item row="2" column="0">
-        <widget class="QPushButton" name="newButton">
-         <property name="text">
-          <string>New game</string>
-         </property>
-        </widget>
-       </item>
-       <item row="1" column="1">
-        <widget class="QLineEdit" name="passwordInput"/>
-       </item>
-       <item row="2" column="1">
-        <widget class="QPushButton" name="joinButton">
-         <property name="text">
-          <string>Join game</string>
-         </property>
-        </widget>
-       </item>
-       <item row="1" column="0">
-        <widget class="QLabel" name="label_3">
-         <property name="text">
-          <string>Password</string>
-         </property>
-        </widget>
-       </item>
        <item row="0" column="0" colspan="2">
         <widget class="QTableWidget" name="sessionsTable">
          <property name="editTriggers">
@@ -134,6 +110,20 @@
            <string>Private</string>
           </property>
          </column>
+        </widget>
+       </item>
+       <item row="1" column="0">
+        <widget class="QPushButton" name="newButton">
+         <property name="text">
+          <string>New game</string>
+         </property>
+        </widget>
+       </item>
+       <item row="1" column="1">
+        <widget class="QPushButton" name="joinButton">
+         <property name="text">
+          <string>Join game</string>
+         </property>
         </widget>
        </item>
       </layout>
